PSARC/2008/336 Extended VTOC 6708609 Support booting on disks > 1TB (phase 1) 6672960 cmlb: default minor nodes for disks > 1TB is not EFI/GPT style 6727240 When a fat32 partition created in the part of the disk after 1TB, block I/O and umount fail

+++ usr/src/grub/grub-0.95/stage2/stage1_5.c
@@ -19,10 +19,10 @@
 
 #include "shared.h"
 
-static int saved_sector = -1;
+static unsigned int saved_sector = (unsigned int)-1;
 
 static void
-disk_read_savesect_func (int sector, int offset, int length)
+disk_read_savesect_func (unsigned int sector, int offset, int length)
 {
   saved_sector = sector;
 }
@@ -45,7 +45,7 @@ cmain (void)
       disk_read_hook = NULL;
 
       /* Sanity check: catch an internal error.  */
-      if (saved_sector == -1)
+      if (saved_sector == (unsigned int)-1)
 	{
 	  grub_printf ("internal error: the second sector of Stage 2 is unknown.");
 	  stop ();
